+++ backend/package_Josh.json
@@ -1,0 +1,56 @@
+{
+	"name": "cryptochain",
+	"version": "1.0.0",
+	"description": "",
+	"main": "index.js",
+	"scripts": {
+		"test": "jest --watchAll --verbose false",
+		"start": "npm run build-client && node index.js",
+		"dev": "npm run dev-client & npm run start-redis && nodemon index.js",
+		"start-redis": "/redis/redis-server redis/redis.windows-service.conf",
+		"node_1": "npm run dev-client & npm run start-redis && nodemon --watch src -e js src/NetworkNode.js 5555 http://localhost:5555",
+		"node_2": "npm run dev-client & npm run start-redis && nodemon --watch src -e js src/NetworkNode.js 5556 http://localhost:5556",
+		"node_3": "npm run dev-client & npm run start-redis && nodemon --watch src -e js src/NetworkNode.js 5557 http://localhost:5557",
+		"node_4": "npm run dev-client & npm run start-redis && nodemon --watch src -e js src/NetworkNode.js 5558 http://localhost:5558",
+		"node_5": "npm run dev-client & npm run start-redis && nodemon --watch src -e js src/NetworkNode.js 5559 http://localhost:5559",
+		"dev-peer": "cross-env GENERATE_PEER_PORT='true' nodemon && index.js",
+		"build-client": "npm run clean && parcel build client/src/index.html --out-dir client/dist",
+		"dev-client": "npm run clean && parcel client/src/index.html --out-dir client/dist",
+		"clean": "rm -rf .cache client/dist"
+	},
+	"jest": {
+		"testEnvironment": "node"
+	},
+	"keywords": [],
+	"author": "",
+	"license": "ISC",
+	"devDependencies": {
+		"cross-env": "^5.2.0",
+		"jest": "^23.6.0",
+		"nodemon": "^1.18.4"
+	},
+	"dependencies": {
+		"babel-core": "^6.26.3",
+		"babel-plugin-transform-class-properties": "^6.24.1",
+		"babel-plugin-transform-object-rest-spread": "^6.26.0",
+		"babel-preset-env": "^1.7.0",
+		"babel-preset-react": "^6.24.1",
+		"body-parser": "^1.18.3",
+		"cors": "^2.8.5",
+		"dotenv": "^16.0.3",
+		"elliptic": "^6.4.1",
+		"express": "^4.16.3",
+		"express-list-endpoints": "^6.0.0",
+		"hex-to-binary": "^1.0.1",
+		"history": "^4.7.2",
+		"http-status-codes": "^2.2.0",
+		"pubnub": "^4.21.6",
+		"react": "^16.6.0",
+		"react-bootstrap": "^0.32.4",
+		"react-dom": "^16.6.0",
+		"react-router-dom": "^4.3.1",
+		"redis": "^2.8.0",
+		"request": "^2.88.0",
+		"uuid": "^3.3.2"
+	}
+}
