+++ frontend/package.json
@@ -14,9 +14,11 @@
 		"@headlessui/react": "^1.7.14",
 		"@heroicons/react": "^2.0.17",
 		"bootstrap": "^5.2.3",
+		"ethers": "^6.3.0",
 		"react": "^18.2.0",
 		"react-bootstrap": "^2.7.3",
-		"react-dom": "^18.2.0"
+		"react-dom": "^18.2.0",
+		"react-router-dom": "^6.10.0"
 	},
 	"devDependencies": {
 		"@types/react": "^18.0.28",
