+++ frontend/src/pages/Explorer.jsx
@@ -1,0 +1,14 @@
+import "../../custom.css";
+import React from "react";
+
+function Explorer() {
+    return (
+      <div>
+        <br />
+        <h1>Explorer</h1>
+      </div>
+    );
+   
+  }
+  
+  export default Explorer;
