+++ frontend/src/pages/Wallet.jsx
@@ -1,0 +1,14 @@
+import "../../custom.css";
+import React from "react";
+
+function Wallet() {
+    return (
+      <div>
+        <br />
+        <h1>Wallet</h1>
+      </div>
+    );
+   
+  }
+  
+  export default Wallet;
